+++ src/EURID/Helper/EppHelper.php
@@ -5,10 +5,12 @@ declare(strict_types=1);
 namespace Upmind\ProvisionProviders\DomainNames\EURID\Helper;
 
 use Carbon\Carbon;
+use Upmind\ProvisionBase\Exception\ProvisionFunctionError;
 use Upmind\ProvisionBase\Helper;
 use Upmind\ProvisionProviders\DomainNames\Data\ContactParams;
 use Upmind\ProvisionProviders\DomainNames\Data\DacDomain;
 use Upmind\ProvisionProviders\DomainNames\Data\DomainNotification;
+use Upmind\ProvisionProviders\DomainNames\Data\Enums\ContactType;
 use Upmind\ProvisionProviders\DomainNames\Helper\Utils;
 use Metaregistrar\EPP\authEppInfoDomainRequest;
 use Metaregistrar\EPP\eppCheckDomainRequest;
@@ -37,6 +39,13 @@ use Upmind\ProvisionProviders\DomainNames\EURID\EppExtension\euridEppTransferDom
 
 class EppHelper
 {
+    /**
+     * Epp contact type constants
+     */
+    private const CONTACT_TYPE_REGISTRANT = 'registrant';
+    private const CONTACT_TYPE_TECH = 'tech';
+    private const CONTACT_TYPE_BILLING = 'billing';
+
     protected EppConnection $connection;
     protected Configuration $configuration;
     private array $lockedStatuses = [
@@ -171,7 +180,7 @@ class EppHelper
         $techId = $response->getDomainContact(eppContactHandle::CONTACT_TYPE_TECH);
         $adminId = $response->getDomainContact(eppContactHandle::CONTACT_TYPE_ADMIN);
 
-        $statuses = $response->getDomainStatuses() ?? [];
+        $statuses = $this->statusesToStrings($response->getDomainStatuses() ?? []);
 
         if ($response->getQuarantined()) {
             $statuses[] = 'quarantined';
@@ -193,7 +202,7 @@ class EppHelper
             'id' => $response->getDomainId(),
             'domain' => $response->getDomainName(),
             'statuses' => $statuses,
-            'locked' => boolval(array_intersect($this->lockedStatuses, $response->getDomainStatuses() ?? [])),
+            'locked' => boolval(array_intersect($this->lockedStatuses, $this->statusesToStrings($response->getDomainStatuses() ?? []))),
             'registrant' => $registrantId ? $this->getContactInfo($registrantId) : null,
             'billing' => $billingId ? $this->getContactInfo($billingId) : null,
             'tech' => $techId ? $this->getContactInfo($techId) : null,
@@ -203,6 +212,22 @@ class EppHelper
             'updated_at' => Utils::formatDate($response->getDomainUpdateDate() ?: $response->getDomainCreateDate()),
             'expires_at' => Utils::formatDate($response->getDomainExpirationDate()),
         ];
+    }
+
+    /**
+     * @param string[]|\Metaregistrar\EPP\eppStatus[] $statuses
+     *
+     * @return string[]
+     */
+    protected function statusesToStrings(array $statuses): array
+    {
+        return array_map(function ($status) {
+            if ($status instanceof \Metaregistrar\EPP\eppStatus) {
+                return $status->getStatusname();
+            }
+
+            return (string)$status;
+        }, $statuses);
     }
 
     /**
@@ -369,6 +394,65 @@ class EppHelper
     }
 
     /**
+     * @throws \Metaregistrar\EPP\eppException
+     * @throws \Propaganistas\LaravelPhone\Exceptions\NumberParseException
+     */
+    public function updateContact(string $domainName, ContactParams $params, ContactType $contactType): ContactData
+    {
+        // If registrant update,
+        // we need to set it to the domain properties and perform update,
+        // as we only have 1 registrant contact.
+        if ($contactType->equals(ContactType::REGISTRANT())) {
+            return $this->updateRegistrantContact($domainName, $params);
+        }
+
+        // For other contact types (tech, billing),
+        // we need to add the new contact and remove the old one of the same type.
+
+        // Get the existing domain info to find existing contacts
+        $info = new eppInfoDomainRequest(new eppDomain($domainName));
+
+        /** @var \Metaregistrar\EPP\eppInfoDomainResponse $response */
+        $response = $this->connection->request($info);
+
+        // Get existing contacts, and set empty array if none.
+        /** @var \Metaregistrar\EPP\eppContactHandle[] $contacts */
+        $contacts = $response->getDomainContacts() ?? [];
+
+        $providerContactType = $this->getProviderContactTypeValue($contactType);
+
+        // Placeholder for remove info
+        $removeInfo = null;
+
+        foreach ($contacts as $contact) {
+            if ($contact->getContactType() !== $providerContactType) {
+                continue;
+            }
+
+            $removeInfo = $removeInfo ?? new eppDomain($domainName);
+            $removeInfo->addContact(new eppContactHandle($contact->getContactHandle(), $providerContactType));
+        }
+
+        // Create the contact, regardless which type it will be created in
+        $contactId = $this->createContact($params, $contactType->getValue());
+
+        // Now add info with the new contact
+        $addInfo = new eppDomain($domainName);
+        $addInfo->addContact(new eppContactHandle($contactId, $providerContactType));
+
+        $update = new eppUpdateDomainRequest(
+            new eppDomain($domainName),
+            $addInfo,
+            $removeInfo,
+            null
+        );
+
+        $this->connection->request($update);
+
+        return $this->getContactInfo($contactId);
+    }
+
+    /**
      * @throws \Propaganistas\LaravelPhone\Exceptions\NumberParseException
      */
     private function setUpdateContactParams(ContactParams $params): eppContact
@@ -520,5 +604,22 @@ class EppHelper
         // Process Response
         /** @var eppTransferResponse */
         return $this->connection->request($transferRequest);
+    }
+
+    /**
+     * @throws \Upmind\ProvisionBase\Exception\ProvisionFunctionError
+     */
+    private function getProviderContactTypeValue(ContactType $contactType): string
+    {
+        switch ($contactType) {
+            case $contactType->equals(ContactType::REGISTRANT()):
+                return self::CONTACT_TYPE_REGISTRANT;
+            case $contactType->equals(ContactType::BILLING()):
+                return self::CONTACT_TYPE_BILLING;
+            case $contactType->equals(ContactType::TECH()):
+                return self::CONTACT_TYPE_TECH;
+            default:
+                throw ProvisionFunctionError::create('Invalid contact type: ' . $contactType->getValue());
+        }
     }
 }