+++ src/OpenSRS/Provider.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Upmind\ProvisionProviders\DomainNames\OpenSRS;
 
 use Carbon\Carbon;
+use Carbon\CarbonImmutable;
 use Carbon\CarbonInterface;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\TransferException;
@@ -13,6 +14,7 @@ use GuzzleHttp\Promise\Utils as PromiseUtils;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 use Throwable;
+use UnexpectedValueException;
 use Upmind\ProvisionBase\Exception\ProvisionFunctionError;
 use Upmind\ProvisionBase\Helper;
 use Upmind\ProvisionBase\Provider\Contract\ProviderInterface;
@@ -26,6 +28,7 @@ use Upmind\ProvisionProviders\DomainNames\Data\DacParams;
 use Upmind\ProvisionProviders\DomainNames\Data\DacResult;
 use Upmind\ProvisionProviders\DomainNames\Data\DomainInfoParams;
 use Upmind\ProvisionProviders\DomainNames\Data\DomainResult;
+use Upmind\ProvisionProviders\DomainNames\Data\Enums\ContactType;
 use Upmind\ProvisionProviders\DomainNames\Data\EppCodeResult;
 use Upmind\ProvisionProviders\DomainNames\Data\EppParams;
 use Upmind\ProvisionProviders\DomainNames\Data\IpsTagParams;
@@ -38,6 +41,7 @@ use Upmind\ProvisionProviders\DomainNames\Data\LockParams;
 use Upmind\ProvisionProviders\DomainNames\Data\PollParams;
 use Upmind\ProvisionProviders\DomainNames\Data\PollResult;
 use Upmind\ProvisionProviders\DomainNames\Data\TransferParams;
+use Upmind\ProvisionProviders\DomainNames\Data\UpdateContactParams;
 use Upmind\ProvisionProviders\DomainNames\Data\UpdateDomainContactParams;
 use Upmind\ProvisionProviders\DomainNames\Data\UpdateNameserversParams;
 use Upmind\ProvisionProviders\DomainNames\Data\VerificationStatusParams;
@@ -123,7 +127,7 @@ class Provider extends DomainNames implements ProviderInterface
                     'tlds' => $tlds,
                 ]
             ], [
-                'timeout' => 10, // Set a reduced timeout for the request
+                'timeout' => 15, // Set a reduced timeout for the request
             ]);
 
         $dacDomains = [];
@@ -545,10 +549,7 @@ class Provider extends DomainNames implements ProviderInterface
                 ]
             ]);
         } catch (ProvisionFunctionError $e) {
-            if (
-                Str::contains($e->getMessage(), 'Authentication Error')
-                && !Str::contains($e->getMessage(), 'Registrar API Authentication Error')
-            ) {
+            if (Str::contains($e->getMessage(), 'Registrant (end-user) authentication error')) {
                 // this actually means domain not found
                 $this->errorResult('Domain name not found', $e->getData(), $e->getDebug(), $e);
             }
@@ -773,7 +774,70 @@ class Provider extends DomainNames implements ProviderInterface
      */
     public function updateRegistrantContact(UpdateDomainContactParams $params): ContactResult
     {
-        return $this->updateContact($params->sld, $params->tld, $params->contact, OpenSrsApi::CONTACT_TYPE_REGISTRANT);
+        return $this->updateContact(UpdateContactParams::create([
+            'sld' => $params->sld,
+            'tld' => $params->tld,
+            'contact' => $params->contact,
+            'contact_type' => ContactType::REGISTRANT,
+        ]));
+    }
+
+    /**
+     * @throws \Upmind\ProvisionBase\Exception\ProvisionFunctionError
+     * @throws \Throwable
+     */
+    public function updateContact(UpdateContactParams $params): ContactResult
+    {
+        try {
+            $contactType = $params->getContactTypeEnum();
+        } catch (UnexpectedValueException $ex) {
+            $this->errorResult('Invalid contact type: ' . $params->contact_type);
+        }
+
+        $type = $this->getProviderContactTypeValue($contactType);
+
+        try {
+            $nameParts = OpenSrsApi::getNameParts($params->contact->name ?? $params->contact->organisation);
+
+            $this->api()->makeRequest([
+                'action' => 'UPDATE_CONTACTS',
+                'object' => 'DOMAIN',
+                'protocol' => 'XCP',
+                'attributes' => [
+                    'domain' => Utils::getDomain($params->sld, $params->tld),
+                    'types' => [$type],
+                    'contact_set' => [
+                        $type => [
+                            'country' => Utils::normalizeCountryCode($params->contact->country_code),
+                            'state' => Utils::stateNameToCode($params->contact->country_code, $params->contact->state),
+                            'org_name' => $params->contact->organisation,
+                            'phone' => Utils::internationalPhoneToEpp($params->contact->phone),
+                            'postal_code' => $params->contact->postcode,
+                            'city' => $params->contact->city,
+                            'email' => $params->contact->email,
+                            'address1' => $params->contact->address1,
+                            'first_name' => $nameParts['firstName'],
+                            'last_name' => $nameParts['lastName'],
+                        ]
+                    ]
+                ]
+            ]);
+
+            return ContactResult::create([
+                'contact_id' => strtolower($type),
+                'name' => $params->contact->name,
+                'email' => $params->contact->email,
+                'phone' => $params->contact->phone,
+                'organisation' => $params->contact->organisation,
+                'address1' => $params->contact->address1,
+                'city' => $params->contact->city,
+                'postcode' => $params->contact->postcode,
+                'country_code' => $params->contact->country_code,
+                'state' => Utils::stateNameToCode($params->contact->country_code, $params->contact->state),
+            ])->setMessage('Contact details updated');
+        } catch (Throwable $e) {
+            $this->handleError($e, $params);
+        }
     }
 
     /**
@@ -984,57 +1048,6 @@ class Provider extends DomainNames implements ProviderInterface
     }
 
     /**
-     * @throws \GuzzleHttp\Exception\GuzzleException
-     * @throws \Throwable
-     * @throws \Upmind\ProvisionBase\Exception\ProvisionFunctionError
-     */
-    private function updateContact(string $sld, string $tld, ContactParams $params, string $type): ContactResult
-    {
-        try {
-            $nameParts = OpenSrsApi::getNameParts($params->name ?? $params->organisation);
-
-            $updateContactRaw = $this->api()->makeRequest([
-                'action' => 'UPDATE_CONTACTS',
-                'object' => 'DOMAIN',
-                'protocol' => 'XCP',
-                'attributes' => [
-                    'domain' => Utils::getDomain($sld, $tld),
-                    'types' => [$type],
-                    'contact_set' => [
-                        $type => [
-                            'country' => Utils::normalizeCountryCode($params->country_code),
-                            'state' => Utils::stateNameToCode($params->country_code, $params->state),
-                            'org_name' => $params->organisation,
-                            'phone' => Utils::internationalPhoneToEpp($params->phone),
-                            'postal_code' => $params->postcode,
-                            'city' => $params->city,
-                            'email' => $params->email,
-                            'address1' => $params->address1,
-                            'first_name' => $nameParts['firstName'],
-                            'last_name' => $nameParts['lastName'],
-                        ]
-                    ]
-                ]
-            ]);
-
-            return ContactResult::create([
-                'contact_id' => strtolower($type),
-                'name' => $params->name,
-                'email' => $params->email,
-                'phone' => $params->phone,
-                'organisation' => $params->organisation,
-                'address1' => $params->address1,
-                'city' => $params->city,
-                'postcode' => $params->postcode,
-                'country_code' => $params->country_code,
-                'state' => Utils::stateNameToCode($params->country_code, $params->state),
-            ])->setMessage('Contact details updated');
-        } catch (Throwable $e) {
-            $this->handleError($e, $params);
-        }
-    }
-
-    /**
      * @param \Throwable $e Encountered error
      * @param DataSet|mixed[] $params
      *
@@ -1082,9 +1095,95 @@ class Provider extends DomainNames implements ProviderInterface
      */
     public function getStatus(DomainInfoParams $params): StatusResult
     {
-        return StatusResult::create()
-            ->setStatus(StatusResult::STATUS_UNKNOWN)
-            ->setExpiresAt(null)
-            ->setRawStatuses(null);
+        $domainName = Utils::getDomain(Arr::get($params, 'sld'), Arr::get($params, 'tld'));
+
+        try {
+            $domainRaw = $this->api()->makeRequest([
+                'action' => 'GET',
+                'object' => 'DOMAIN',
+                'protocol' => 'XCP',
+                'attributes' => [
+                    'domain' => $domainName,
+                    'type' => 'all_info',
+                    'clean_ca_subset' => 1,
+                    // 'active_contacts_only' => 1
+                ]
+            ]);
+
+            $expiryDate = CarbonImmutable::parse($domainRaw['attributes']['expiredate']);
+
+            if ($expiryDate->isPast()) {
+                return StatusResult::create()
+                    ->setStatus(StatusResult::STATUS_EXPIRED)
+                    ->setExpiresAt($expiryDate);
+            }
+
+            return StatusResult::create()
+                ->setStatus(StatusResult::STATUS_ACTIVE)
+                ->setExpiresAt($expiryDate);
+        } catch (ProvisionFunctionError $e) {
+            $result = $this->api()->makeRequest([
+                'action' => 'GET_DELETED_DOMAINS',
+                'object' => 'DOMAIN',
+                'protocol' => 'XCP',
+                'attributes' => [
+                    'domain' => $domainName,
+                ]
+            ]);
+
+            $deletedDomain = $result['attributes']['del_domains'][0] ?? null;
+
+            if (isset($deletedDomain)) {
+                switch ($deletedDomain['reason']) {
+                    case 'Transfered':
+                        return StatusResult::create()
+                            ->setStatus(StatusResult::STATUS_TRANSFERRED_AWAY)
+                            ->setExpiresAt(null)
+                            ->setExtra([
+                                'delete_info' => $deletedDomain,
+                            ]);
+                    case 'Expired':
+                        return StatusResult::create()
+                            ->setStatus(StatusResult::STATUS_EXPIRED)
+                            ->setExpiresAt(Carbon::createFromTimeString($deletedDomain['expiredate_epoch']))
+                            ->setExtra([
+                                'delete_info' => $deletedDomain,
+                            ]);
+                    default:
+                        return StatusResult::create()
+                            ->setStatus(StatusResult::STATUS_CANCELLED)
+                            ->setExpiresAt(null)
+                            ->setExtra([
+                                'delete_info' => $deletedDomain,
+                            ]);
+                }
+            }
+
+            if (Str::contains($e->getMessage(), 'Registrant (end-user) authentication error')) {
+                // this actually means domain not found
+                $this->errorResult('Domain name not found', $e->getData(), $e->getDebug(), $e);
+            }
+
+            throw $e;
+        }
+    }
+
+    /**
+     * @throws \Upmind\ProvisionBase\Exception\ProvisionFunctionError
+     */
+    private function getProviderContactTypeValue(ContactType $contactType): string
+    {
+        switch ($contactType) {
+            case $contactType->equals(ContactType::REGISTRANT()):
+                return OpenSrsApi::CONTACT_TYPE_REGISTRANT;
+            case $contactType->equals(ContactType::ADMIN()):
+                return OpenSrsApi::CONTACT_TYPE_ADMIN;
+            case $contactType->equals(ContactType::BILLING()):
+                return OpenSrsApi::CONTACT_TYPE_BILLING;
+            case $contactType->equals(ContactType::TECH()):
+                return OpenSrsApi::CONTACT_TYPE_TECH;
+            default:
+                throw ProvisionFunctionError::create('Invalid contact type: ' . $contactType->getValue());
+        }
     }
 }